requirements.txt
fastapi[standard]==0.115.12
pyserial==3.5
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
cloudinary==1.38.0
python-multipart==0.0.20
pillow==11.2.1
